+++ activemodel/lib/active_model.rb
@@ -35,6 +35,7 @@ module ActiveModel
   autoload :Dirty
   autoload :EachValidator, 'active_model/validator'
   autoload :ForbiddenAttributesProtection
+  autoload :Identity
   autoload :Lint
   autoload :Model
   autoload :Name, 'active_model/naming'