+++ activerecord/lib/active_record/core.rb
@@ -7,6 +7,8 @@ module ActiveRecord
     extend ActiveSupport::Concern
 
     included do
+      include ActiveModel::Identity
+
       ##
       # :singleton-method:
       #
@@ -346,33 +348,6 @@ module ActiveRecord
       coder['raw_attributes'] = attributes_before_type_cast
       coder['attributes'] = @attributes
       coder['new_record'] = new_record?
-    end
-
-    # Returns true if +comparison_object+ is the same exact object, or +comparison_object+
-    # is of the same type and +self+ has an ID and it is equal to +comparison_object.id+.
-    #
-    # Note that new records are different from any other record by definition, unless the
-    # other record is the receiver itself. Besides, if you fetch existing records with
-    # +select+ and leave the ID out, you're on your own, this predicate will return false.
-    #
-    # Note also that destroying a record preserves its ID in the model instance, so deleted
-    # models are still comparable.
-    def ==(comparison_object)
-      super ||
-        comparison_object.instance_of?(self.class) &&
-        !id.nil? &&
-        comparison_object.id == id
-    end
-    alias :eql? :==
-
-    # Delegates to id in order to allow two records of the same type and id to work with something like:
-    #   [ Person.find(1), Person.find(2), Person.find(3) ] & [ Person.find(1), Person.find(4) ] # => [ Person.find(1) ]
-    def hash
-      if id
-        id.hash
-      else
-        super
-      end
     end
 
     # Clone and freeze the attributes hash such that associations are still